+++ examples/src/test/scala/Http4sExample.scala
@@ -14,22 +14,30 @@
  * limitations under the License.
  */
 
+import scala.concurrent.ExecutionContext
+import scala.concurrent.duration._
+
+import cats.Parallel
 import cats.data.NonEmptyList
-import cats.effect.IO
+import cats.effect._
 import cats.instances.list._
-import cats.syntax.traverse._
+import cats.syntax.all._
+import cats.temp.par._
+
 import io.circe._
 import io.circe.generic.semiauto._
+
 import org.http4s.circe._
 import org.http4s.client.blaze._
-import org.scalatest.{AsyncWordSpec, Matchers}
+import org.scalatest.{Matchers, WordSpec}
 
-import scala.concurrent.{ExecutionContext, Future}
 import fetch._
-import fetch.implicits._
 
-class Http4sExample extends AsyncWordSpec with Matchers {
-  implicit override def executionContext: ExecutionContext = ExecutionContext.Implicits.global
+class Http4sExample extends WordSpec with Matchers {
+  implicit val executionContext = ExecutionContext.Implicits.global
+
+  implicit val t: Timer[IO]         = IO.timer(executionContext)
+  implicit val cs: ContextShift[IO] = IO.contextShift(executionContext)
 
   // in this example we are fetching users and their posts via http using http4s
   // the demo api is https://jsonplaceholder.typicode.com/
@@ -51,106 +59,100 @@ class Http4sExample extends AsyncWordSpec with Matchers {
 
   // http4s client which is used by the datasources
 
-  val client = PooledHttp1Client[IO]()
+  def client[F[_]: ConcurrentEffect] =
+    Http1Client[F](
+      BlazeClientConfig.defaultConfig.copy(
+        responseHeaderTimeout = 30.seconds // high timeout because jsonplaceholder takes a while to respond
+      ))
 
   // a DataSource that can fetch Users with their UserId.
 
-  implicit val userDS = new DataSource[UserId, User] {
+  object Users extends DataSource[UserId, User] {
     override def name = "UserH4s"
-    override def fetchOne(id: UserId): Query[Option[User]] =
-      Query.async { (ok, fail) =>
-        fetchById(id).unsafeRunAsync(_.fold(fail, ok))
-      }
-    override def fetchMany(ids: NonEmptyList[UserId]): Query[Map[UserId, User]] =
-      Query.async { (ok, fail) =>
-        fetchByIds(ids)
-          .map(users => users.map(user => user.id -> user).toMap)
-          .unsafeRunAsync(_.fold(fail, ok))
-      }
 
-    // fetchById and fetchByIds would probably be defined in some other module
-
-    def fetchById(id: UserId): IO[Option[User]] = {
+    override def fetch[F[_]: ConcurrentEffect: Par](id: UserId): F[Option[User]] = {
       val url = s"https://jsonplaceholder.typicode.com/users?id=${id.id}"
-      client.expect(url)(jsonOf[IO, List[User]]).map(_.headOption)
+      client[F] >>= ((c) => c.expect(url)(jsonOf[F, List[User]]).map(_.headOption))
     }
 
-    def fetchByIds(ids: NonEmptyList[UserId]): IO[List[User]] = {
+    override def batch[F[_]: ConcurrentEffect: Par](
+        ids: NonEmptyList[UserId]
+    ): F[Map[UserId, User]] = {
       val filterIds = ids.map("id=" + _.id).toList.mkString("&")
       val url       = s"https://jsonplaceholder.typicode.com/users?$filterIds"
-      client.expect(url)(jsonOf[IO, List[User]])
+      val io        = client[F] >>= ((c) => c.expect(url)(jsonOf[F, List[User]]))
+      io.map(users => users.map(user => user.id -> user).toMap)
     }
   }
 
   // a datasource that can fetch all the Posts using a UserId
 
-  implicit val postsForUserDS = new DataSource[UserId, List[Post]] {
+  object Posts extends DataSource[UserId, List[Post]] {
     override def name = "PostH4s"
-    override def fetchOne(id: UserId): Query[Option[List[Post]]] =
-      Query.async { (ok, fail) =>
-        fetchById(id).map(Option.apply).unsafeRunAsync(_.fold(fail, ok))
-      }
-    override def fetchMany(ids: NonEmptyList[UserId]): Query[Map[UserId, List[Post]]] =
-      Query.async { (ok, fail) =>
-        fetchByIds(ids).unsafeRunAsync(_.fold(fail, ok))
-      }
-
-    def fetchById(id: UserId): IO[List[Post]] = {
+    override def fetch[F[_]: ConcurrentEffect: Par](id: UserId): F[Option[List[Post]]] = {
       val url = s"https://jsonplaceholder.typicode.com/posts?userId=${id.id}"
-      client.expect(url)(jsonOf[IO, List[Post]])
+      client[F] >>= ((c) => c.expect(url)(jsonOf[F, List[Post]]).map(Option.apply))
     }
 
-    def fetchByIds(ids: NonEmptyList[UserId]): IO[Map[UserId, List[Post]]] = {
+    override def batch[F[_]: ConcurrentEffect: Par](
+        ids: NonEmptyList[UserId]): F[Map[UserId, List[Post]]] = {
       val filterIds = ids.map("userId=" + _.id).toList.mkString("&")
       val url       = s"https://jsonplaceholder.typicode.com/posts?$filterIds"
-      client.expect(url)(jsonOf[IO, List[Post]]).map(_.groupBy(_.userId).toMap)
+      client[F] >>= ((c) => c.expect(url)(jsonOf[F, List[Post]]).map(_.groupBy(_.userId).toMap))
     }
   }
 
   // some helper methods to create Fetches
 
-  def user(id: UserId): Fetch[User]               = Fetch(id)
-  def postsForUser(id: UserId): Fetch[List[Post]] = Fetch(id)
+  def user[F[_]: ConcurrentEffect](id: UserId): Fetch[F, User] =
+    Fetch(id, Users)
+
+  def postsForUser[F[_]: ConcurrentEffect](id: UserId): Fetch[F, List[Post]] =
+    Fetch(id, Posts)
 
   "We can fetch one user" in {
-    val fetch: Fetch[User]            = user(UserId(1))
-    val fut: Future[(FetchEnv, User)] = Fetch.runFetch[Future](fetch)
-    fut.map {
-      case (env, user) =>
-        println(user)
-        env.rounds.size shouldEqual 1
-    }
+    def fetch[F[_]: ConcurrentEffect] = user(UserId(1))
+
+    val io: IO[(Env, User)] = Fetch.runEnv[IO](fetch)
+
+    val (env, result) = io.unsafeRunSync
+
+    println(result)
+    env.rounds.size shouldEqual 1
   }
 
   "We can fetch multiple users in parallel" in {
-    val fetch: Fetch[List[User]] = List(1, 2, 3).traverse(i => user(UserId(i)))
-    val fut                      = Fetch.runFetch[Future](fetch)
-    fut.map {
-      case (env, users) =>
-        users.foreach(println)
-        env.rounds.size shouldEqual 1
-    }
+    def fetch[F[_]: ConcurrentEffect]: Fetch[F, List[User]] =
+      List(1, 2, 3).traverse(i => user(UserId(i)))
+
+    val io = Fetch.runEnv[IO](fetch)
+
+    val (env, result) = io.unsafeRunSync
+
+    result.foreach(println)
+    env.rounds.size shouldEqual 1
   }
 
   "We can fetch multiple users with their posts" in {
-    val fetch: Fetch[List[(User, List[Post])]] =
+    def fetch[F[_]: ConcurrentEffect]: Fetch[F, List[(User, List[Post])]] =
       for {
-        users <- List(UserId(1), UserId(2)).traverse(user)
+        users <- List(UserId(1), UserId(2)).traverse(user[F])
         usersWithPosts <- users.traverse { user =>
           postsForUser(user.id).map(posts => (user, posts))
         }
       } yield usersWithPosts
-    val fut = Fetch.runFetch[Future](fetch)
-    fut.map {
-      case (env, userPosts) =>
-        userPosts
-          .map {
-            case (user, posts) =>
-              s"${user.username} has ${posts.size} posts"
-          }
-          .foreach(println)
-        env.rounds.size shouldEqual 2
-    }
+
+    val io = Fetch.runEnv[IO](fetch)
+
+    val (env, results) = io.unsafeRunSync
+
+    results
+      .map {
+        case (user, posts) =>
+          s"${user.username} has ${posts.size} posts"
+      }
+      .foreach(println)
+    env.rounds.size shouldEqual 2
   }
 
 }